+++ rpm-repository/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 Released on ReleaseDate.
 
+### Changed
+
+* Handle float timestamp in `repomd.xml` (but saved rounded `u64` value).
+
 ## 0.2.0
 
 Released on 2023-11-03.